MY-46 exp search and confirm

+++ src/main/java/asia/atmonline/myriskservice/web/bureau/ccris/dto/confirm_entity/request/ExperianCCRISEntityRequestBody.java
@@ -16,7 +16,7 @@ import org.apache.commons.lang3.builder.ToStringStyle;
 public class ExperianCCRISEntityRequestBody {
 
   private String ProductType;
-  private Integer CRefId;
+  private Long CRefId;
   private String EntityKey;
   private String MobileNo;
   private String EmailAddress;
